DSDEGP-1261: jenkins test cleanups
- merge jenkins test code into other test code
- remove unused files, configs
- disable verbose output from pytest
