--- clio-integration-test/src/test/resources/docker-compose-jenkins/elasticsearch-compose.yml
@@ -1,20 +1,0 @@
-version: '2'
-
-services:
-  elasticsearch:
-    image: ${ELASTICSEARCH_DOCKER_TAG}
-    environment:
-      cluster.name: docker-cluster
-      bootstrap.memory_lock: 'true'
-      ES_JAVA_OPTS: -Xms512m -Xmx512m
-      xpack.security.enabled: 'false'
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-      nofile:
-        soft: 65536
-        hard: 65536
-    mem_limit: 1g
-    cap_add:
-      - IPC_LOCK
