+++ clio-integration-test/src/test/resources/docker-compose/docker-compose-jenkins.yml
